Revert "Filtrado en front search-job"
