+++ src/components/hero.tsx
@@ -1,70 +1,27 @@
 import * as React from "react";
-import { useEffect, useState } from "react";
 import "./../sass/components/_hero.scss";
 import ButtonComponent from "./shared/atom/button";
-import { Txtfield, DropdownMenu, DropdownItem } from "./shared/styled";
-import { Category } from "../interfaces/Category";
-import { Position } from "../interfaces/Position";
-import { useNavigate } from "react-router-dom";
-import { getCategory, getPosition } from "../util/publication.service";
+import {
+  Txtfield,
+  DropdownMenu,
+  DropdownItem,
+  BtnPrimary,
+} from "./shared/styled";
+
+import imgCover from "./../assets/imgcover.png"
+import {useNavigate} from "react-router-dom";
 
 export default function Hero() {
-  const [categorys, setCategorys] = useState<Category[]>([]);
-  const [positions, setPositions] = useState<Position[]>([]);
-  
-  const [position, setPosition] = React.useState({
-		idCategory: '',
-		namePosition: '',
-		description: '',
-	});
-  //Get all categorys
-  const getAllCategories = async () => {
-    const c = await getCategory();
-    console.log({ c });
-    if (c.status) setCategorys(c?.data);
-  };
 
-  const handleCategory = async (e: any) => {
-    const value = e.target.value;
-    setPositions([]);
-    if (value !== "0") {
-      setPosition({ ...position, idCategory: value });
-      const p = await getPosition({ idCategory: value });
-      setPositions(p.data);
+    const navigate = useNavigate();
+    /**
+     * Handle redirects to user register form.
+     */
+    const handleRedirect = () => {
+        navigate("/register")
     }
-  };
-
-  useEffect(() => {
-    (async () => {
-      await getAllCategories();
-    })();
-  }, []);
-
-  const navigate = useNavigate();
-  /**
-   * Handle redirects to user register form.
-   */
-  const handleRedirect = () => {
-    navigate("/register");
-  };
   return (
     <React.Fragment>
-      <div className="secFull">
-        <section className="HeroComponent">
-          <aside className="cover">
-            <article>
-              <h1>
-                <b>
-                  EL INICIO DE TU <br /> PRÓXIMO TRABAJO
-                </b>
-              </h1>
-              <div className="btnContainer">
-                <ButtonComponent family="hero1" label="Quiero trabajar" />
-                <ButtonComponent family="hero2" label="Quiero contratar" />
-              </div>
-            </article>
-          </aside>
-        </section>
       <div className="secFull"> 
       <section className="HeroComponent">
         <aside className="cover">
@@ -81,28 +38,15 @@ export default function Hero() {
       </section>
       </div>
       <aside className="filterSection">
-        <DropdownMenu name="category" onChange={handleCategory}>
-          <DropdownItem value="_none">Seleccionar Categoría</DropdownItem>
-
-          {categorys.map((category: Category) => (
-            <DropdownItem key={category._id} value={category._id}>
-              {category.nameCategory}
-            </DropdownItem>
-          ))}
-        </DropdownMenu>
-
-        <DropdownMenu name="position" >
-							<DropdownItem>Eliga el puesto</DropdownItem>
-							{positions.map((e: any) => (
-								<DropdownItem key={e._id} value={e._id}>
-								{e.namePosition}
-								</DropdownItem>
-								))}
-						</DropdownMenu>
-
-        <Txtfield placeholder="Ubicación" />
-        <ButtonComponent family="primary" label="FILTRAR" />
-      </aside>
+          <Txtfield placeholder="Especialidad" />
+          <Txtfield placeholder="Ubicación" />
+          <DropdownMenu>
+            <DropdownItem>Seleccionar Categoría</DropdownItem>
+            <DropdownItem>Diseño</DropdownItem>
+            <DropdownItem>Ingeniería</DropdownItem>
+          </DropdownMenu>
+          <ButtonComponent family="primary" label="FILTRAR" />
+        </aside>
     </React.Fragment>
   );
 }
