+++ src/sass/components/_listPost.scss
@@ -4,7 +4,7 @@
     display: flex;
     flex-direction: column;
     justify-content: center;
-    margin: 0.5em 0;
+    margin: 80px 0;
     text-align: center;
     padding: 0px 12%;
     @include tablet-l {
