+++ src/types/post_job.d.ts
@@ -64,8 +64,6 @@ export interface PostJob {
 	title?: string;
 	updateDate?: number;
 	listSkills?: any[];
-	idCategory:string;
-	idPosition:string;
 	postulants?: number;
 }
